feat(mobile): use system browser for OAuth login instead of in-app WebView
Replace the in-app WebView OAuth flow with the system browser
(ASWebAuthenticationSession on iOS, Chrome Custom Tabs on Android) via
nativescript-inappbrowser. authFlow.loginWithOAuth now calls
InAppBrowser.openAuth(authorizeUrl, REDIRECT_URI) and exchanges the
returned code for tokens.

This is the RFC 8252 recommended pattern for native apps and resolves a
string of in-WebView issues on Android: the custom com.frappe.mail://
redirect scheme couldn't be intercepted by the bare WebView
(ERR_UNKNOWN_URL_SCHEME), and presenting the WebView as a modal corrupted
frame navigation on iOS (post-login white screen).

Required changes:
- Android: NativeScriptActivity needs launchMode="singleTask" so the
  OAuth redirect deep link resumes the existing activity (onNewIntent)
  instead of spawning a blank duplicate (which forced an app restart).
- Pin nativescript-inappbrowser to 3.2.0 — the 3.3.0 "latest" is a broken
  publish that ships type defs and native artifacts but no JavaScript.

Removes the OAuthWebView component, the native WebViewClient interop, and
the modal machinery. Navigation is driven explicitly from the page
handlers; logout returns to the landing page.
