+++ mobile/app/pages/LandingPage.vue
@@ -1,6 +1,5 @@
 <template>
-	<Page>
-		<ActionBar :title="__('Frappe Mail')" />
+	<Page actionBarHidden="true">
 		<ScrollView>
 			<StackLayout class="p-6">
 				<Label
@@ -47,28 +46,28 @@
 					:text="__('YOUR SITES')"
 					class="text-ink-gray-5 mb-2 mt-5 text-sm font-semibold"
 				/>
-				<StackLayout
+				<GridLayout
 					v-for="s in site.sites"
 					:key="s.url"
+					columns="*, auto"
 					class="bg-surface-gray-1 mb-2 rounded-lg p-4"
-					@tap="selectAndLogin(s.url, s.client_id)"
 				>
-					<GridLayout columns="*, auto">
-						<StackLayout col="0">
-							<Label
-								:text="s.app_name || s.sitename"
-								class="text-ink-gray-9 text-base font-semibold"
-							/>
-							<Label :text="s.url" class="text-ink-gray-5 text-sm" textWrap="true" />
-						</StackLayout>
+					<!-- Tap target is only the content column, so tapping ✕ doesn't
+					     bubble up into a login (NativeScript tap gestures propagate). -->
+					<StackLayout col="0" @tap="selectAndLogin(s.url, s.client_id)">
 						<Label
-							col="1"
-							text="✕"
-							class="text-ink-gray-4 p-2"
-							@tap="site.removeSite(s.url)"
+							:text="s.app_name || s.sitename"
+							class="text-ink-gray-9 text-base font-semibold"
 						/>
-					</GridLayout>
-				</StackLayout>
+						<Label :text="s.url" class="text-ink-gray-5 text-sm" textWrap="true" />
+					</StackLayout>
+					<Label
+						col="1"
+						text="✕"
+						class="text-ink-gray-4 p-2"
+						@tap="site.removeSite(s.url)"
+					/>
+				</GridLayout>
 
 				<ActivityIndicator v-if="busy" :busy="busy" class="mt-4" />
 			</StackLayout>
@@ -78,11 +77,13 @@
 
 <script setup lang="ts">
 import { ref } from 'vue'
+import { $navigateTo } from 'nativescript-vue'
 
 import { useApi } from '@/utils/api'
 import { loadTranslations } from '@/utils/translation'
 import { sessionStore } from '@/stores/session'
 import { siteStore } from '@/stores/site'
+import AppShell from '@/pages/AppShell.vue'
 
 import type { SiteInfo } from '@mail/types'
 
@@ -134,6 +135,7 @@ async function addSite() {
 	}
 }
 
+// Open the system-browser OAuth flow; on success, move on to the app shell.
 async function selectAndLogin(url: string, clientId: string) {
 	error.value = ''
 	site.setActiveSite(url)
@@ -141,6 +143,7 @@ async function selectAndLogin(url: string, clientId: string) {
 	try {
 		await session.login(url, clientId)
 		await loadTranslations()
+		$navigateTo(AppShell, { clearHistory: true })
 	} catch (e) {
 		error.value = messageOf(e, __('Login failed'))
 	} finally {
