+++ mobile/package.json
@@ -17,6 +17,7 @@
     "@nativescript/core": "~8.9.0",
     "@nativescript/secure-storage": "^4.0.1",
     "dayjs": "^1.11.11",
+    "nativescript-inappbrowser": "3.2.0",
     "nativescript-vue": "^3.0.0",
     "pinia": "^2.1.7",
     "vue": "^3.5"
